+++ src/components/WowerlayReusables/index.ts
@@ -1,21 +1,35 @@
 import { PropType } from 'vue';
 
 export interface WowerlayBaseProps {
-  align?: 'auto' | 'top' | 'bottom';
+  position?: 'left' | 'right' | 'top' | 'bottom';
+  verticalGap: number;
+  horizontalGap: number;
   canLeaveViewport?: boolean;
-  noFollow?: boolean;
-  target: any;
+  fixed?: boolean;
+  target?: HTMLElement;
   tag?: string;
 }
 
 export const wowerlayBaseProps = {
   target: {
     required: true,
-    type: null
+    type: null as unknown as PropType<WowerlayBaseProps['target']>
   },
-  align: {
-    default: 'auto',
-    type: String as PropType<WowerlayBaseProps['align']>
+  fixed: {
+    default: false,
+    type: Boolean as PropType<WowerlayBaseProps['fixed']>
+  },
+  position: {
+    default: 'bottom',
+    type: String as PropType<WowerlayBaseProps['position']>
+  },
+  verticalGap: {
+    default: 0,
+    type: Number as PropType<WowerlayBaseProps['verticalGap']>
+  },
+  horizontalGap: {
+    default: 0,
+    type: Number as PropType<WowerlayBaseProps['horizontalGap']>
   },
   canLeaveViewport: {
     default: false,