+++ CHANGELOG.md
@@ -1,3 +1,11 @@
+## 0.1.0
+
+- `Fixed` prop is added
+- `vertical-gap` and `horizontal-gap` props added. With these props spacing between Wowerlay and target element can be configurable.
+- 'position' prop is added. If set `bottom | right | left | top` Wowerlay will try to be mounted to given direction, if can't it will try opposite side.
+- Reactive `target` prop support is added. Now wowerlay will react when `target` element changes and will follow it.
+- Demo page added to Readme
+
 ## 0.0.1
 
 - First Ready to use build.